Full cleanup, UI using constant messages

+++ src/com/thiefg/asteroids/subobjects/DistortedPolygon.java
@@ -10,7 +10,7 @@ public class DistortedPolygon extends Polygon {
 	super(location, sides, radius);
 	distances = new ArrayList<>();
 	for (int i = 0; i < sides; i++) {
-	    distances.add(new Double((double) radius * Math.random() * (2-.5) + .5));
+	    distances.add(new Double(radius * Math.random() * (2-.5) + .5));
 	}
     }
 
@@ -19,7 +19,7 @@ public class DistortedPolygon extends Polygon {
 	super(location, sides, radius, rotation);
 	distances = new ArrayList<>();
 	for (int i = 0; i < sides; i++) {
-	    distances.add(new Double((double) radius * Math.random() * (2-.5) + .5));
+	    distances.add(new Double(radius * Math.random() * (2-.5) + .5));
 	}
     }
 
@@ -28,17 +28,17 @@ public class DistortedPolygon extends Polygon {
 	super(location, sides, radius, rotation);
 	distances = new ArrayList<>();
 	for (int i = 0; i < sides; i++) {
-	    distances.add(new Double((double) radius * Math.random() * (maxChange-minChange) + minChange));
+	    distances.add(new Double(radius * Math.random() * (maxChange-minChange) + minChange));
 	}
     }
 
     @Override
     public void update() {
 	setPoints(new ArrayList<Vector2d>());
-	double innerAngle = 360d / (double) super.getSides();
+	double innerAngle = 360d / super.getSides();
 	for (int i = 0; i < super.getSides(); i++) {
 	    getPoints().add(
-		    new Vector2d(super.getLocation(), innerAngle * (double) i
+		    new Vector2d(super.getLocation(), innerAngle * i
 			    + (innerAngle / 2d) + super.getRotation(), distances.get(i).doubleValue()));
 	}
     }
